Finished Map
On click - Sets marker at click location with a bound popup to it
containing the address of that point. Upon clicking on the marker, the
popup will go away leaving only the marker (putting the popup in its
closed state). The marker will stay in its location and is clickable to
make the popup come up again (in its open state).